+++ Patches/PunManagerPatch.cs
@@ -1,23 +1,24 @@
 ﻿using HarmonyLib;
-using Steamworks;
+using UnityEngine;
 
 namespace RepoAdminMenu.Patches {
 
     [HarmonyPatch(typeof(PunManager))]
     internal class PunManagerPatch {
 
-        [HarmonyPatch("UpdateHealthRightAway")]
+        [HarmonyPatch(nameof(PunManager.UpdateHealthRightAway))]
         [HarmonyPrefix]
-        private static bool Prefix_UpdateHealthRightAway(ref string playerName) {
-            PlayerAvatar playerAvatar = SemiFunc.PlayerAvatarGetFromSteamID(playerName);
+        private static bool Prefix_UpdateHealthRightAway(ref string _steamID) {
+            PlayerAvatar playerAvatar = SemiFunc.PlayerAvatarGetFromSteamID(_steamID);
             if (playerAvatar == SemiFunc.PlayerAvatarLocal()) {
-                playerAvatar.playerHealth.maxHealth = 100 + (StatsManager.instance.playerUpgradeHealth[playerName] * 20);
-                playerAvatar.playerHealth.Heal(playerAvatar.playerHealth.maxHealth - playerAvatar.playerHealth.health, false);
+                playerAvatar.playerHealth.maxHealth = 100 + (StatsManager.instance.playerUpgradeHealth[_steamID] * 20);
+                playerAvatar.playerHealth.health = playerAvatar.playerHealth.maxHealth - 1;
+                playerAvatar.playerHealth.Heal(1, false);
             }
             return false;
         }
 
-        [HarmonyPatch("UpdateEnergyRightAway")]
+        [HarmonyPatch(nameof(PunManager.UpdateEnergyRightAway))]
         [HarmonyPrefix]
         private static bool Prefix_UpdateEnergyRightAway(ref string _steamID) {
             if (SemiFunc.PlayerAvatarGetFromSteamID(_steamID) == SemiFunc.PlayerAvatarLocal()) {
@@ -27,7 +28,7 @@ namespace RepoAdminMenu.Patches {
             return false;
         }
 
-        [HarmonyPatch("UpdateExtraJumpRightAway")]
+        [HarmonyPatch(nameof(PunManager.UpdateExtraJumpRightAway))]
         [HarmonyPrefix]
         private static bool Prefix_UpdateExtraJumpRightAway(ref string _steamID) {
             if (SemiFunc.PlayerAvatarGetFromSteamID(_steamID) == SemiFunc.PlayerAvatarLocal()) {
@@ -36,7 +37,7 @@ namespace RepoAdminMenu.Patches {
             return false;
         }
 
-        [HarmonyPatch("UpdateMapPlayerCountRightAway")]
+        [HarmonyPatch(nameof(PunManager.UpdateMapPlayerCountRightAway))]
         [HarmonyPrefix]
         private static bool UpdateMapPlayerCountRightAway(ref string _steamID) {
             PlayerAvatar playerAvatar = SemiFunc.PlayerAvatarGetFromSteamID(_steamID);
@@ -46,24 +47,24 @@ namespace RepoAdminMenu.Patches {
             return false;
         }
 
-        [HarmonyPatch("UpdateTumbleLaunchRightAway")]
+        [HarmonyPatch(nameof(PunManager.UpdateTumbleLaunchRightAway))]
         [HarmonyPrefix]
         private static bool Prefix_UpdateTumbleLaunchRightAway(ref string _steamID) {
             SemiFunc.PlayerAvatarGetFromSteamID(_steamID).tumble.tumbleLaunch = StatsManager.instance.playerUpgradeLaunch[_steamID];
             return false;
         }
 
-        [HarmonyPatch("UpdateSprintSpeedRightAway")]
+        [HarmonyPatch(nameof(PunManager.UpdateSprintSpeedRightAway))]
         [HarmonyPrefix]
         private static bool Prefix_UpdateSprintSpeedRightAway(ref string _steamID) {
             if (SemiFunc.PlayerAvatarGetFromSteamID(_steamID) == SemiFunc.PlayerAvatarLocal()) {
-                PlayerController.instance.SprintSpeed = StatsManager.instance.playerUpgradeSpeed[_steamID];
-                PlayerController.instance.SprintSpeedUpgrades = StatsManager.instance.playerUpgradeSpeed[_steamID];
+                PlayerController.instance.SprintSpeed = 5f + (float) StatsManager.instance.playerUpgradeSpeed[_steamID];
+                PlayerController.instance.SprintSpeedUpgrades = (float) StatsManager.instance.playerUpgradeSpeed[_steamID];
             }
             return false;
         }
 
-        [HarmonyPatch("UpdateGrabStrengthRightAway")]
+        [HarmonyPatch(nameof(PunManager.UpdateGrabStrengthRightAway))]
         [HarmonyPrefix]
         private static bool Prefix_UpdateGrabStrengthRightAway(ref string _steamID) {
             PlayerAvatar playerAvatar = SemiFunc.PlayerAvatarGetFromSteamID(_steamID);
@@ -73,7 +74,7 @@ namespace RepoAdminMenu.Patches {
             return false;
         }
 
-        [HarmonyPatch("UpdateThrowStrengthRightAway")]
+        [HarmonyPatch(nameof(PunManager.UpdateThrowStrengthRightAway))]
         [HarmonyPrefix]
         private static bool Prefix_UpdateThrowStrengthRightAway(ref string _steamID) {
             PlayerAvatar playerAvatar = SemiFunc.PlayerAvatarGetFromSteamID(_steamID);
@@ -83,12 +84,32 @@ namespace RepoAdminMenu.Patches {
             return false;
         }
 
-        [HarmonyPatch("UpdateGrabRangeRightAway")]
+        [HarmonyPatch(nameof(PunManager.UpdateGrabRangeRightAway))]
         [HarmonyPrefix]
         private static bool Prefix_UpdateGrabRangeRightAway(ref string _steamID) {
             PlayerAvatar playerAvatar = SemiFunc.PlayerAvatarGetFromSteamID(_steamID);
             if (playerAvatar) {
                 playerAvatar.physGrabber.grabRange = 4 + (StatsManager.instance.playerUpgradeRange[_steamID] * 1f);
+            }
+            return false;
+        }
+
+        [HarmonyPatch(nameof(PunManager.UpdateCrouchRestRightAway))]
+        [HarmonyPrefix]
+        private static bool Prefix_UpdateCrouchRestRightAway(ref string _steamID) {
+            PlayerAvatar playerAvatar = SemiFunc.PlayerAvatarGetFromSteamID(_steamID);
+            if (playerAvatar) {
+                playerAvatar.upgradeCrouchRest = StatsManager.instance.playerUpgradeCrouchRest[_steamID];
+            }
+            return false;
+        }
+
+        [HarmonyPatch(nameof(PunManager.UpdateTumbleWingsRightAway))]
+        [HarmonyPrefix]
+        private static bool Prefix_UpdateTumbleWingsRightAway(ref string _steamID) {
+            PlayerAvatar playerAvatar = SemiFunc.PlayerAvatarGetFromSteamID(_steamID);
+            if (playerAvatar) {
+                playerAvatar.upgradeTumbleWings = StatsManager.instance.playerUpgradeTumbleWings[_steamID];
             }
             return false;
         }