+++ RepoAdminMenu.cs
@@ -1,7 +1,6 @@
 ﻿using BepInEx;
 using BepInEx.Logging;
 using HarmonyLib;
-using RepoAdminMenu.Utils;
 
 namespace RepoAdminMenu {
 
@@ -12,7 +11,7 @@ namespace RepoAdminMenu {
 
         private const string mod_guid = "proferabg.REPO.RepoAdminMenu";
         private const string mod_name = "Repo Admin Menu";
-        private const string mod_version = "1.0.10";
+        private const string mod_version = "1.0.11";
 
         private static RepoAdminMenu _plugin;